# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated products",
                "operationId": "ec5306a3c5913c4991e8b5bf30c441f2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store Product",
                "operationId": "9351d26c781ab39f2668abdb2be6108d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "product_selling_price",
                                    "product_buying_price",
                                    "unit_id",
                                    "product_stock"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "product_selling_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "product_buying_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "unit_id": {
                                        "type": "integer"
                                    },
                                    "product_stock": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStoreResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/insert": {
            "post": {
                "tags": [
                    "TransactionsMaster"
                ],
                "summary": "Create transaction and update product stock",
                "operationId": "91a78d57c0df4d210f754b2ecbad6503",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionInsertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionInsertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PaginatedProductResponse": {
                "allOf": [
                    {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/PaginatedResponseBase"
                    }
                ]
            },
            "PaginatedResponseBase": {
                "properties": {
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "page": {
                                    "type": "integer",
                                    "nullable": true
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "product_price": {
                        "type": "string",
                        "example": "77614.00"
                    },
                    "product_stock": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "TransactionItem": {
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 1000
                    },
                    "qty": {
                        "type": "integer",
                        "example": 4
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "TransactionMaster": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 13
                    },
                    "transaction_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-25T15:19:57.982423Z"
                    },
                    "transaction_type": {
                        "$ref": "#/components/schemas/TransactionType"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransactionItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-25T15:19:57.000000Z"
                    }
                },
                "type": "object"
            },
            "TransactionType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "type_name": {
                        "type": "string",
                        "example": "Pengeluaran"
                    },
                    "is_decreasing": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Unit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "unit_name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "is_big_unit": {
                        "type": "string",
                        "example": "77614.00"
                    },
                    "smallest_unit_id": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "TransactionInsertRequest": {
                "required": [
                    "type",
                    "items"
                ],
                "properties": {
                    "type": {
                        "description": "Must exist in transaction_types.typename",
                        "type": "string",
                        "example": "OUT"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransactionItemRequest"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
            "TransactionItemRequest": {
                "required": [
                    "productId",
                    "productPrice",
                    "qty"
                ],
                "properties": {
                    "productId": {
                        "type": "integer",
                        "example": 1
                    },
                    "productPrice": {
                        "type": "integer",
                        "example": 1000
                    },
                    "qty": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 5
                    }
                },
                "type": "object"
            },
            "ProductStoreResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Product"
                    },
                    {
                        "properties": {
                            "unit": {
                                "$ref": "#/components/schemas/Unit"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "TransactionInsertResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Transaction created successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TransactionMaster"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "TransactionsMaster",
            "description": "TransactionsMaster"
        }
    ]
}